+++ src/components/LibraryListFormFields/index.tsx
@@ -1,10 +1,8 @@
-import { Heading } from "@nypl/design-system-react-components";
+import { Heading, Select } from "@nypl/design-system-react-components";
 import { useTranslation } from "next-i18next";
 import { useState } from "react";
 import { useFormContext } from "react-hook-form";
-import Autosuggest from "react-autosuggest";
 
-import FormField from "../FormField";
 import useFormDataContext from "../../context/FormDataContext";
 import { findLibraryName } from "../../utils/formDataUtils";
 import { LibraryListObject } from "../../interfaces";
@@ -30,99 +28,38 @@ const LibraryListForm = ({ libraryList = [] }: LibraryListFormProps) => {
     ? findLibraryName(formValues?.homeLibraryCode)
     : "";
   const [value, setValue] = useState(defaultValue);
-  const [suggestions, setSuggestions] = useState([]);
   const { register } = useFormContext();
 
-  const onChange = (event, { newValue }) => {
-    setValue(newValue);
+  const onChange = (event) => {
+    setValue(event.target.value);
   };
-  /**
-   * getSuggestions
-   * This function tells react-autosuggest how to filter the suggestions. In this
-   * case, if the user input is found anywhere in any of the libraries' name,
-   * return the library. This is to "search" even if the user doesn't start with
-   * the library's first name.
-   */
-  const getSuggestions = (value: string, list: LibraryListObject[]) => {
-    const inputValue = value.trim().toLowerCase();
-    const findInString = (l) =>
-      l.label.toLowerCase().indexOf(inputValue) !== -1;
-    return inputValue.length === 0 ? [] : list.filter(findInString);
-  };
-  /**
-   * getSuggestionValue
-   * All we need is the label of the library object as the value.
-   */
-  const getSuggestionValue = (suggestion: LibraryListObject) =>
-    suggestion.label;
-  /**
-   * renderSuggestion
-   * How to display each suggestion which gets rendered in a list item element.
-   */
-  const renderSuggestion = (suggestion: LibraryListObject) => (
-    <div>{suggestion.label}</div>
-  );
-  // Autosuggest will call this function every time suggestions need to be
-  // updated.
-  const onSuggestionsFetchRequested = ({ value }) =>
-    setSuggestions(getSuggestions(value, libraryList));
-  // Clear suggestions.
-  const onSuggestionsClearRequested = () => setSuggestions([]);
-  // Autosuggest will pass through all these props to the input.
-
   const inputProps = {
-    placeholder: `${t("account.library.placeholder")}`,
     value,
     onChange,
     // Pass in the `react-hook-form` register function so it can handle this
     // form element's state for us.
     ref: register(),
   };
-  /**
-   * renderInputComponent
-   * This is the custom component we want to render for the form field.
-   */
-  const renderInputComponent = (inputProps) => {
-    // According to axe, this is a bad aria label,
-    // so let's remove it.
-    delete inputProps["aria-autocomplete"];
-
-    return (
-      <FormField
-        id="librarylist-autosuggest"
-        label={t("account.library.selectLibrary")}
-        name="homeLibraryCode"
-        isRequired={false}
-        attributes={{ ...inputProps }}
-      />
-    );
-  };
-  /**
-   * renderSuggestionsContainer
-   * To solve an accessibility issue, we render the autosuggest container
-   * with an `aria-label`.
-   */
-  const renderSuggestionsContainer = ({ containerProps, children }) => (
-    <div {...containerProps} aria-label="List of library name suggestions.">
-      {children}
-    </div>
-  );
 
   return (
     <div className={styles.container}>
       <Heading level="three">{t("account.library.title")}</Heading>
       <p>{t("account.library.description.part1")}</p>
       <p>{t("account.library.description.part2")}</p>
-      <Autosuggest
-        suggestions={suggestions}
-        onSuggestionsFetchRequested={onSuggestionsFetchRequested}
-        onSuggestionsClearRequested={onSuggestionsClearRequested}
-        getSuggestionValue={getSuggestionValue}
-        renderSuggestion={renderSuggestion}
-        inputProps={inputProps}
-        renderInputComponent={renderInputComponent}
-        renderSuggestionsContainer={renderSuggestionsContainer}
-      />
+      <Select
+        placeholder="Please select"
+        id="librarylist-select"
+        labelText={t("account.library.selectLibrary")}
+        name="homeLibraryCode"
+        isRequired={false}
+        {...inputProps}
+      >
+        {libraryList.map(({ value, label }, i) => (
+          <option key={i} value={value}>
+            {label}
+          </option>
+        ))}
+      </Select>
     </div>
   );
 };