+++ src/utils/api.ts
@@ -409,9 +409,6 @@ export async function callPatronAPI(
         error: err.message,
       };
     }
-    if (err.response?.data.detail.includes("PIN is trivial"))
-      err.response.data.detail =
-        "Password cannot contain consecutively repeating characters three or more times, e.g. aaaatf54 or repeating a pattern, e.g. abcabcab";
     if (status === 401 || status === 403) {
       serverError = { type: "internal" };
     }
@@ -445,11 +442,12 @@ export async function createPatron(
   appObj = app
 ) {
   const data = req.body;
+  // default to e branch
+  if (!data.homeLibraryCode) data.homeLibraryCode = "eb";
   const csrfTokenValid = validateCsrfToken(req);
   if (!csrfTokenValid) {
     return invalidCsrfResponse(res);
   }
-
   try {
     const results = await callPatronAPI(data, createPatronUrl, appObj);
     res.status(results.status).json(results);