+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 ## CHANGE LOG
 
+### 1.2.4 
+- replace home library free text input with select drop down[SCC-4660](https://newyorkpubliclibrary.atlassian.net/browse/SCC-4660)
+- revert verify password match to original state
+- update auth url [SCC-4657](https://newyorkpubliclibrary.atlassian.net/browse/SCC-4657)
+- replace SimplyE label with EBranch [SCC-4659](https://newyorkpubliclibrary.atlassian.net/browse/SCC-4659)
+
+### 1.2.3 Fix PIN translation errors
+- use password instructions for "PIN is trivial" error translation
+- add id's to facilitate QA testing in other languages
+
 ### 1.2.2 CSRF and resubmission bugs
 - remove csrf token from reducer state, instead pass in as prop
 - add more descriptive server logs for csrf token debugging